CK/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
pandas==2.2.3
openpyxl==3.1.50
numpy==1.24.4
urllib3==2.1.0
